sclang: remove unused code and move opcode details to new file
